upgrade dependencies & introduce github ci

+++ store/atime_linux.go
@@ -1,3 +1,4 @@
+//go:build linux
 // +build linux
 
 package store
@@ -9,7 +10,7 @@ import (
 )
 
 func timespecToTime(ts syscall.Timespec) time.Time {
-	return time.Unix(int64(ts.Sec), int64(ts.Nsec))
+	return time.Unix(ts.Sec, ts.Nsec)
 }
 
 func atime(fi os.FileInfo) time.Time {
